requirements.txt
asttokens==2.4.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
decorator==5.1.1
dill==0.3.7
exceptiongroup==1.2.0
executing==2.0.1
idna==3.6
ipython==8.19.0
ipywidgets==8.1.1
jedi==0.19.1
jupyterlab-widgets==3.0.9
matplotlib-inline==0.1.6
numpy==1.26.3
pandas==2.1.4
parso==0.8.3
praw==7.7.1
prawcore==2.4.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
scipy==1.11.4
six==1.16.0
stack-data==0.6.3
traitlets==5.14.1
tzdata==2023.4
update-checker==0.18.0
urllib3==2.1.0
wcwidth==0.2.12
websocket-client==1.7.0
widgetsnbextension==4.0.9
xmltodict==0.13.0
